+++ web/src/types/ApiResponse.ts
@@ -3,3 +3,12 @@ export interface ApiResponse<T> {
   succeeded: boolean;
   errors: string[];
 }
+
+export class CustomAPIError<T> extends Error {
+  public originalError: T;
+  constructor(message, originalError) {
+    super(message);
+    this.name = 'CustomAPIError';
+    this.originalError = originalError; // Store the original Axios error
+  }
+}
