+++ web/src/utils/utils.ts
@@ -318,12 +318,3 @@ export const parseQueryStringToString = (
 export const isPositiveInteger = (value) => {
   return _.isNumber(value) && value > 0;
 };
-
-export class CustomAPIError<T> extends Error {
-  public originalError: T;
-  constructor(message, originalError) {
-    super(message);
-    this.name = 'CustomAPIError';
-    this.originalError = originalError; // Store the original Axios error
-  }
-}
